File upload improved. Design improved. New modules were implemented. Notification functionality was impelemented.

+++ AspMVCMeeting/Angular/MeetingLineIndependent/Controller.js
@@ -1,4 +1,26 @@
-﻿appIndep.controller("IndepCntrl", function ($scope, $http, $filter, $q, NgTableParams, upload, IndepService) {
+﻿appIndep.controller("IndepCntrl", function ($scope, $document, $http, $filter, $q, NgTableParams, upload, IndepService) {
+
+    var clearDepWatch = null;
+
+    var getData = IndepService.GetAllByColumnName("MTL_SCODE1", "MEETING_LINES");
+    getData.then(function (emp) {
+        $scope.lineScopes1 = emp.data;
+    });
+
+    var getData = IndepService.GetAllByColumnName("MTL_SCODE2", "MEETING_LINES");
+    getData.then(function (emp) {
+        $scope.lineScopes2 = emp.data;
+    });
+
+    var getData = IndepService.GetAllByColumnName("MTL_SCODE3", "MEETING_LINES");
+    getData.then(function (emp) {
+        $scope.lineScopes3 = emp.data;
+    });
+
+    var getData = IndepService.GetAllByColumnName("MTL_SCODE4", "MEETING_LINES");
+    getData.then(function (emp) {
+        $scope.lineScopes4 = emp.data;
+    });
 
     //BEGIN Master LINES files create
 
@@ -84,28 +106,28 @@
 
     GetLineAll();
 
-    function GetLineAll() {
-
-        var getData = IndepService.GetLinesAll(null);
+    function GetLineAll(notClose) {
+        var getData = IndepService.GetLinesAll(angular.element("#MEETING_LINES_MTL_MT_REF").val());
         getData.then(function (emp) {
             $scope.data = emp.data;
+
             $scope.tableParams = new NgTableParams({
                 page: 1, // show first page
                 total: 1, // value less than count hide pagination
-                count: 5 // count per page
+                count: 12 // count per page
             }, { counts: [], dataset: $scope.data });
-
 
 
             $scope.checkboxes = { 'checked': false, items: {} };
             angular.forEach($scope.data, function (item) {
-                $scope.checkboxes.items[item.MEETING_LINES.ID] = false;
+                if (item.MEETING_LINES.MTL_STS === 4)
+                    $scope.checkboxes.items[item.MEETING_LINES.ID] = false;
             });
 
             // watch for check all checkbox
             $scope.$watch('checkboxes.checked', function (value) {
                 angular.forEach($scope.data, function (item) {
-                    if (angular.isDefined(item.MEETING_LINES.ID)) {
+                    if (angular.isDefined(item.MEETING_LINES.ID) && item.MEETING_LINES.MTL_STS === 4) {
                         $scope.checkboxes.items[item.MEETING_LINES.ID] = value;
                     }
                 });
@@ -116,18 +138,31 @@
                 if (!$scope.data) {
                     return;
                 }
-                var checked = 0, unchecked = 0,
-                    total = $scope.data.length;
+                var checked = 0, unchecked = 0, total = 0;
+
                 angular.forEach($scope.data, function (item) {
-                    checked += ($scope.checkboxes.items[item.MEETING_LINES.ID]) || 0;
-                    unchecked += (!$scope.checkboxes.items[item.MEETING_LINES.ID]) || 0;
+                    if (item.MEETING_LINES.MTL_STS === 4) total += 1;
+                });
+
+                angular.forEach($scope.data, function (item) {
+                    if (item.MEETING_LINES.MTL_STS === 4) {
+                        checked += ($scope.checkboxes.items[item.MEETING_LINES.ID]) || 0;
+                        unchecked += (!$scope.checkboxes.items[item.MEETING_LINES.ID]) || 0;
+                    }
                 });
                 if ((unchecked == 0) || (checked == 0)) {
                     $scope.checkboxes.checked = (checked == total);
+                    if (total == 0) $scope.checkboxes.checked = false;
                 }
                 // grayed checkbox
                 angular.element($("#select_all")).prop("indeterminate", (checked != 0 && unchecked != 0));
             }, true);
+
+            if (!angular.isDefined(notClose))
+                $scope.CancelAddEdit();
+
+            if (emp.data.length === 0)
+                $scope.AddLineDiv();
 
         }, function (response) {
             alert('Error in getting records');
@@ -163,13 +198,16 @@
     }
 
     $scope.editLine = function (line, isEdit) {
+        if (clearDepWatch != null) clearDepWatch();
         $scope.fileCreate = false;
         $scope.lineFiles = '';
         $scope.resultLine = '';
+        $scope.$watch('line.MEETING_LINES.MTL_RESPONSIBLE', function () { });
         var getData = IndepService.getLine(line.MEETING_LINES.ID);
         getData.then(function (emp) {
             $scope.line = emp.data;
             $scope.Action = "Update";
+
             for (var keyName in $scope.line.MEETING_LINES) {
                 var key = keyName;
                 var value = $scope.line.MEETING_LINES[keyName];
@@ -187,7 +225,7 @@
 
             $scope.fileEdit = true;
             GetAllLineFiles(line.MEETING_LINES.ID);
-            
+
             angular.element("#lineId").val(line.MEETING_LINES.ID);
             angular.element('textarea').removeAttr('style');
             angular.element('#MEETING_LINES_MTL_TAGS').tagsinput('removeAll');
@@ -202,7 +240,17 @@
                 angular.element('#MeetingLinesDiv').find('input, textarea, select').attr('disabled', 'disabled');
                 angular.element('#btnAddLine').hide();
             }
-            
+
+            clearDepWatch = $scope.$watch('line.MEETING_LINES.MTL_RESPONSIBLE', function (newValue, oldValue) {
+                if (newValue !== oldValue) {
+                    //alert(newValue + " " + oldValue);
+                    findDepByAccount(newValue);
+                }
+            });
+
+            angular.element("#AddCopyLine").css("display", "none");
+            $scope.editClass = {};
+            $scope.editClass[line.MEETING_LINES.ID] = 'active';
             angular.element("#MeetingLinesDiv").slideDown();
 
         },
@@ -212,19 +260,42 @@
     }
 
     $scope.AddLineDiv = function () {
-        $scope.line.MEETING_LINES.MTL_START_DATE = convertDate($scope.line.MEETING_LINES.MTL_START_DATE);
+        if (clearDepWatch !== null) clearDepWatch();
+        angular.element('#MeetingLinesDiv').find('input, textarea, select').removeAttr('disabled');
         angular.element('#btnAddLine').show();
         $scope.fileEdit = false;
         $scope.lineFiles = '';
         $scope.resultLine = '';
         $scope.line = "";
-        $scope.line = { "MEETING_LINES": { "MTL_MT_REF": angular.element("#MEETING_LINES_MTL_MT_REF").val() } };
+        $scope.line = {
+            "MEETING_LINES": {
+                "MTL_MT_REF": angular.element("#MEETING_LINES_MTL_MT_REF").val(), "MTL_START_DATE": new Date(),
+                "MTL_DECISION_TYPE": angular.element("#MEETING_MASTER_MT_TYPE").val()
+            }
+        };
         $scope.Action = "Add";
         $scope.fileCreate = true;
         angular.element('textarea').removeAttr('style');
         angular.element('#MEETING_LINES_MTL_TAGS').tagsinput('removeAll');
+        //$scope.line.MEETING_LINES.MTL_START_DATE = convertDate($scope.line.MEETING_LINES.MTL_START_DATE);
+
+        clearDepWatch = $scope.$watch('line.MEETING_LINES.MTL_RESPONSIBLE', function (newValue) {
+            //alert(newValue);
+            findDepByAccount(newValue);
+        });
+
+        angular.element("#AddCopyLine").css("display", "");
         angular.element("#MeetingLinesDiv").slideDown();
 
+    }
+
+    $scope.CancelAddEdit = function () {
+        $scope.editClass = {};
+        $scope.Action = null;
+        angular.element("#AddCopyLine").css("display", "none");
+        angular.element("#MeetingLinesDiv").slideUp('fast', function () {
+            $scope.line = null;
+        });
     }
 
     $scope.AddUpdateLine = function () {
@@ -233,37 +304,63 @@
         if ($scope.line.MEETING_LINES.MTL_FINISH_DATE != null)
             $scope.line.MEETING_LINES.MTL_FINISH_DATE = convertDate($scope.line.MEETING_LINES.MTL_FINISH_DATE);
 
-        var getAction = $scope.Action;
-        if (getAction == "Update") {
+        if ($scope.Action === "Update") {
             var getData = IndepService.updateLine($scope.line);
             getData.then(function (msg) {
                 GetLineAll();
-                angular.element("#MeetingLinesDiv").slideUp();
             }, function () {
                 alert('Error in updating record');
             });
         }
-        else {
+        else if ($scope.Action === "Add") {
             var getData = IndepService.AddLine($scope.line);
             getData.then(function (msg) {
                 GetLineAll();
-                angular.element("#MeetingLinesDiv").slideUp();
+                getMasterById(angular.element("#MEETING_LINES_MTL_MT_REF").val());
             }, function () {
                 alert('Error in adding record');
             });
         }
-        debugger;
-        $scope.refresh();
     }
 
-
     $scope.deleteLine = function (line) {
-        var getData = IndepService.DeleteLine(line);
+        if (confirm('Silmek istediyinizden eminmisiniz?')) {
+            var getData = IndepService.DeleteLine(line);
+            getData.then(function (msg) {
+                GetLineAll();
+            }, function () {
+                alert('Error in Deleting Record');
+            });
+        };
+    }
+
+    $scope.copyLine = function (line) {
+        var getData = IndepService.CopyLine(line);
         getData.then(function (msg) {
-            GetLineAll();
+            GetLineAll(true);
+            $scope.editLine(msg.data, true);
         }, function () {
-            alert('Error in Deleting Record');
+            alert('Error in Coping Record');
         });
+    }
+
+    $scope.AddCopyLine = function () {
+        var copiedLine = null;
+        if ($scope.line.MEETING_LINES.MTL_START_DATE != null)
+            $scope.line.MEETING_LINES.MTL_START_DATE = convertDate($scope.line.MEETING_LINES.MTL_START_DATE);
+        if ($scope.line.MEETING_LINES.MTL_FINISH_DATE != null)
+            $scope.line.MEETING_LINES.MTL_FINISH_DATE = convertDate($scope.line.MEETING_LINES.MTL_FINISH_DATE);
+
+        if ($scope.Action === "Add") {
+            var getData = IndepService.AddLine($scope.line);
+            getData.then(function (msg) {
+                GetLineAll(true);
+                $scope.editClass = {};
+
+            }, function () {
+                alert('Error in adding record');
+            });
+        }
     }
 
     $scope.publishLine = function (line) {
@@ -275,5 +372,34 @@
         });
     }
 
+    function findDepByAccount(accountName) {
+        var getData = IndepService.findDepByAccount(accountName);
+        getData.then(function (emp) {
+            $scope.line.MEETING_LINES.MTL_DEPARTMENT = emp.data;
+        },
+        function () {
+            //alert('Error in getting records');
+        });
+
+    }
+
+    $document.bind("keydown", function (event) {
+        if (event.ctrlKey || event.metaKey) {
+            switch (String.fromCharCode(event.which).toLowerCase()) {
+                case 's':
+                    event.preventDefault();
+                    if ($scope.Action != null) {
+                        $scope.AddUpdateLine();
+                        $scope.$apply();
+                    }
+                    break;
+                case 'y':
+                    event.preventDefault();
+                    $scope.AddLineDiv();
+                    $scope.$apply();
+                    break;
+            }
+        }
+    });
 
 });
